+++ ui/src/shared/components/Dygraph.tsx
@@ -459,8 +459,12 @@ class Dygraph extends Component<Props, State> {
   }
 
   private getLabel = (axis: string): string => {
-    const {axes, queries} = this.props
-    const label = getDeep<string>(axes, `${axis}.label`, '')
+    const {axes, labels, queries} = this.props
+
+    // if label comes back as '', use the y axis label from props.labels, if it exists
+    // see https://github.com/influxdata/chronograf/issues/5314
+    const fallbackLabel = labels[1] || ''
+    const label = getDeep<string>(axes, `${axis}.label`, '') || fallbackLabel
     const queryConfig = getDeep(queries, '0.queryConfig', false)
 
     if (label || !queryConfig) {